Refactor Menu/Items
